requirements.txt
discord~=2.2.2
openai~=0.26.5
python-dotenv~=1.0.0
riotwatcher~=3.2.5
quickchart